requirements.txt
streamlit==1.18.1
pandas==1.1.3
base58==2.0.1
numpy==1.19.2
scikit-learn==0.23.2
plotly_express==0.4.0
pandas-profiling==2.10.0
streamlit-pandas-profiling==0.1.1
Jinja2==3.0.3